new update to cleandb

+++ server/src/seeds/cleanDb.ts
@@ -3,14 +3,19 @@ import db from '../config/connection.js';
 
 export default async (modelName: "Question", collectionName: string) => {
   try {
-    let modelExists = await models[modelName]?.db.db.listCollections({
-      name: collectionName
-    }).toArray()
+    const model = models[modelName]; 
+    if (model) { 
+      let modelExists = await model.db.db.listCollections({
+        name: collectionName
+      }).toArray();
 
-    if (modelExists.length) {
-      await db.dropCollection(collectionName);
+      if (modelExists.length) {
+        await db.dropCollection(collectionName);
+      }
+    } else {
+      console.log(`Model ${modelName} not found.`); 
     }
   } catch (err) {
     throw err;
   }
-}
+};
